+++ packages/block-brokers/test/trustless-gateway-sessions.spec.ts
@@ -8,10 +8,11 @@ import { uriToMultiaddr } from '@multiformats/uri-to-multiaddr'
 import { expect } from 'aegir/chai'
 import { CID } from 'multiformats/cid'
 import Sinon from 'sinon'
-import { type StubbedInstance, stubInterface } from 'sinon-ts'
+import { stubInterface } from 'sinon-ts'
 import { createTrustlessGatewaySession } from '../src/trustless-gateway/session.js'
 import type { Routing } from '@helia/interface'
 import type { ComponentLogger } from '@libp2p/interface'
+import type { StubbedInstance } from 'sinon-ts'
 
 interface StubbedTrustlessGatewaySessionComponents {
   logger: ComponentLogger
@@ -108,5 +109,64 @@ describe('trustless-gateway sessions', () => {
 
     await expect(session.retrieve(cid)).to.eventually.deep.equal(block)
     expect(queryProviderSpy.callCount).to.equal(1)
+  })
+
+  it('should end session when signal is aborted', async () => {
+    const cid = CID.parse('bafkreig7p6kzwgg4hp3n7wpnnn3kkjmpzxds5rmwhphyueilbzabvyexvq')
+    const session = createTrustlessGatewaySession(components, {
+      allowInsecure: true,
+      allowLocal: true
+    })
+
+    const queryProviderSpy = Sinon.spy(session, 'queryProvider')
+
+    components.routing.findProviders.returns(async function * () {
+      yield {
+        id: peerIdFromPrivateKey(await generateKeyPair('Ed25519')),
+        multiaddrs: [
+          uriToMultiaddr(process.env.BAD_TRUSTLESS_GATEWAY ?? '')
+        ]
+      }
+    }())
+
+    await expect(session.retrieve(cid, { signal: AbortSignal.timeout(500) })).to.eventually.be.rejected()
+      .with.property('name', 'AbortError')
+    expect(queryProviderSpy.callCount).to.equal(1)
+  })
+
+  it('should not abort the session when the signal is aborted if the block is found', async () => {
+    const cid = CID.parse('bafkreig7p6kzwgg4hp3n7wpnnn3kkjmpzxds5rmwhphyueilbzabvyexvq')
+    const block = Uint8Array.from([0, 1, 2, 0])
+    const session = createTrustlessGatewaySession(components, {
+      allowInsecure: true,
+      allowLocal: true,
+      transformRequestInit: (requestInit) => {
+        requestInit.headers = {
+          // The difference here on my machine is 25ms.. if the difference between finding the block and the signal being aborted is less than 22ms, then the test will fail.
+          delay: '478'
+        }
+        return requestInit
+      }
+    })
+
+    const queryProviderSpy = Sinon.spy(session, 'queryProvider')
+
+    components.routing.findProviders.returns(async function * () {
+      yield {
+        id: peerIdFromPrivateKey(await generateKeyPair('Ed25519')),
+        multiaddrs: [
+          uriToMultiaddr(process.env.BAD_TRUSTLESS_GATEWAY ?? '')
+        ]
+      }
+      yield {
+        id: peerIdFromPrivateKey(await generateKeyPair('Ed25519')),
+        multiaddrs: [
+          uriToMultiaddr(process.env.TRUSTLESS_GATEWAY ?? '')
+        ]
+      }
+    }())
+
+    await expect(session.retrieve(cid, { signal: AbortSignal.timeout(500) })).to.eventually.deep.equal(block)
+    expect(queryProviderSpy.callCount).to.equal(2)
   })
 })