+++ packages/dag-cbor/test/get.spec.ts
@@ -3,7 +3,8 @@
 import { expect } from 'aegir/chai'
 import { MemoryBlockstore } from 'blockstore-core'
 import { identity } from 'multiformats/hashes/identity'
-import { dagCbor, type DAGCBOR } from '../src/index.js'
+import { dagCbor } from '../src/index.js'
+import type { DAGCBOR } from '../src/index.js'
 import type { Blockstore } from 'interface-blockstore'
 import type { CID } from 'multiformats/cid'