+++ packages/interop/src/dag-cbor.spec.ts
@@ -1,11 +1,12 @@
 /* eslint-env mocha */
 
-import { dagCbor, type DAGCBOR, type AddOptions } from '@helia/dag-cbor'
+import { dagCbor } from '@helia/dag-cbor'
 import * as codec from '@ipld/dag-cbor'
 import { expect } from 'aegir/chai'
 import { CID } from 'multiformats/cid'
 import { createHeliaNode } from './fixtures/create-helia.js'
 import { createKuboNode } from './fixtures/create-kubo.js'
+import type { DAGCBOR, AddOptions } from '@helia/dag-cbor'
 import type { HeliaLibp2p } from 'helia'
 import type { KuboNode } from 'ipfsd-ctl'
 import type { AddOptions as KuboAddOptions } from 'kubo-rpc-client'