+++ packages/interop/src/fixtures/create-helia-http.ts
@@ -1,4 +1,5 @@
-import { createHeliaHTTP as createHelia, type HeliaHTTPInit } from '@helia/http'
+import { createHeliaHTTP as createHelia } from '@helia/http'
+import type { HeliaHTTPInit } from '@helia/http'
 import type { Helia } from '@helia/interface'
 
 export async function createHeliaHTTP (init: Partial<HeliaHTTPInit> = {}): Promise<Helia> {