+++ packages/utils/test/abstract-session.spec.ts
@@ -1,4 +1,5 @@
 import { generateKeyPair } from '@libp2p/crypto/keys'
+import { isPeerId } from '@libp2p/interface'
 import { defaultLogger } from '@libp2p/logger'
 import { peerIdFromPrivateKey } from '@libp2p/peer-id'
 import { expect } from 'aegir/chai'
@@ -8,6 +9,7 @@ import { raceSignal } from 'race-signal'
 import Sinon from 'sinon'
 import { AbstractSession } from '../src/abstract-session.js'
 import type { PeerId } from '@libp2p/interface'
+import type { Multiaddr } from '@multiformats/multiaddr'
 import type { ProgressEvent } from 'progress-events'
 
 interface SessionPeer {
@@ -32,6 +34,14 @@ class Session extends AbstractSession<SessionPeer, ProgressEvent> {
 
   equals (a: SessionPeer, b: SessionPeer): boolean {
     return a.id.equals(b.id)
+  }
+
+  async convertToProvider (provider: PeerId | Multiaddr | Multiaddr[]): Promise<SessionPeer | undefined> {
+    if (isPeerId(provider)) {
+      return {
+        id: provider
+      }
+    }
   }
 }
 
@@ -254,5 +264,70 @@ describe('abstract-session', () => {
 
     expect(session.findNewProviders).to.have.property('callCount', 2)
     expect(session.queryProvider).to.have.property('callCount', 1)
+  })
+
+  it('should abort retrieve if the signal is aborted before provider returns block', async () => {
+    const session = new Session()
+
+    const cid = CID.parse('bafybeifaymukvfkyw6xgh4th7tsctiifr4ea2btoznf46y6b2fnvikdczi')
+    const block = Uint8Array.from([0, 1, 2, 3])
+
+    session.findNewProviders.onFirstCall().callsFake(async function * () {
+      yield {
+        id: peerIdFromPrivateKey(await generateKeyPair('Ed25519'))
+      }
+    })
+    session.queryProvider.callsFake(async (cid, provider, options) => {
+      return raceSignal((async () => {
+        await delay(100)
+
+        return block
+      })(), options.signal)
+    })
+
+    await expect(session.retrieve(cid, {
+      signal: AbortSignal.timeout(10)
+    })).to.eventually.be.rejected()
+      .with.property('name', 'AbortError')
+  })
+
+  it('should handle race condition between finding block and abort signal', async () => {
+    const session = new Session()
+
+    const cid = CID.parse('bafybeifaymukvfkyw6xgh4th7tsctiifr4ea2btoznf46y6b2fnvikdczi')
+    const block = Uint8Array.from([0, 1, 2, 3])
+
+    const providers: SessionPeer[] = [{
+      id: peerIdFromPrivateKey(await generateKeyPair('Ed25519'))
+    }, {
+      id: peerIdFromPrivateKey(await generateKeyPair('Ed25519'))
+    }]
+
+    session.findNewProviders.callsFake(async function * () {
+      yield providers[0]
+      yield providers[1]
+    })
+    const abortDelay = 500
+
+    session.queryProvider.withArgs(cid, providers[0]).callsFake(async (_cid, _provider, options) => {
+      return raceSignal((async () => {
+        await delay(abortDelay * 2) // always takes longer than abortDelay
+
+        return Uint8Array.from([0, 1, 2, 3, 4])
+      })(), options.signal)
+    })
+    session.queryProvider.withArgs(cid, providers[1]).callsFake(async (_cid, _provider, options) => {
+      return raceSignal((async () => {
+        await delay(abortDelay - 40)
+
+        return block
+      })(), options.signal)
+    })
+
+    await expect(session.retrieve(cid, {
+      signal: AbortSignal.timeout(abortDelay)
+    })).to.eventually.deep.equal(block)
+
+    expect(session.queryProvider.callCount).to.equal(2)
   })
 })