+++ packages/car/CHANGELOG.md
@@ -1,5 +1,48 @@
 # Changelog
 
+## [4.1.2](https://github.com/ipfs/helia/compare/car-v4.1.1...car-v4.1.2) (2025-05-20)
+
+
+### Dependencies
+
+* update aegir to 47.x.x ([#804](https://github.com/ipfs/helia/issues/804)) ([60fbbc2](https://github.com/ipfs/helia/commit/60fbbc2eb08e023e2eac02ae0e89ed143d715084))
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^5.3.1 to ^5.3.2
+  * devDependencies
+    * @helia/mfs bumped from ^5.0.2 to ^5.0.3
+    * @helia/unixfs bumped from ^5.0.2 to ^5.0.3
+
+## [4.1.1](https://github.com/ipfs/helia/compare/car-v4.1.0...car-v4.1.1) (2025-05-13)
+
+
+### Dependencies
+
+* update all deps ([#792](https://github.com/ipfs/helia/issues/792)) ([d43efc7](https://github.com/ipfs/helia/commit/d43efc7bdfff34071a8e4e22e01f659fbac0b78e))
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^5.3.0 to ^5.3.1
+  * devDependencies
+    * @helia/mfs bumped from ^5.0.1 to ^5.0.2
+    * @helia/unixfs bumped from ^5.0.1 to ^5.0.2
+
+## [4.1.0](https://github.com/ipfs/helia/compare/car-v4.0.4...car-v4.1.0) (2025-05-13)
+
+
+### Features
+
+* car export supports custom dag traversal and export ([#767](https://github.com/ipfs/helia/issues/767)) ([313e2c1](https://github.com/ipfs/helia/commit/313e2c10a71c4f42533c3f077c6720563dd09ce5))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^5.2.1 to ^5.3.0
+  * devDependencies
+    * @helia/mfs bumped from ^5.0.0 to ^5.0.1
+    * @helia/unixfs bumped from ^5.0.0 to ^5.0.1
+
 ## [4.0.4](https://github.com/ipfs/helia/compare/car-v4.0.3...car-v4.0.4) (2025-03-20)